future: final(bug button search)

+++ src/pages/LandingPage.jsx
@@ -1,31 +1,13 @@
 import { useEffect, useState } from "react";
-// import ToDoInput from "../components/ToDoInput";
 import TodoForm from "../components/ToDoForm";
-// import Data from "../data.json";
 import TodoDelete from "../components/ToDoDelete";
 import ToDoSearch from "../components/ToDoSearch";
 import TodoList from "../components/ToDoList";
 import PropTypes from "prop-types";
-// import filterTask from "../components/FilterList";
-
-// localStorage.setItem("userData", JSON.stringify(Data));
-
-// const getLocalStorage = () => {
-//   let Data = localStorage.getItem("Data");
-//   if (Data) {
-//     return (Data = JSON.parse(localStorage.getItem("Data")));
-//   } else {
-//     return [];
-//   }
-// };
 
 function LandingPage({ task, setTask }) {
-  // const [task, setTask] = useState(getLocalStorage());
-  // const [text, setText] = useState("");
   const [show, setShow] = useState(task);
-  // useEffect(() => {
-  //   localStorage.setItem("Data", JSON.stringify(task));
-  // }, [task]);
+  const [searchTask, setSearchTask] = useState("");
 
   //fungsi untuk melihat all, done dan todo
   const handlefilterTask = (showTodo) => {
@@ -38,31 +20,9 @@ function LandingPage({ task, setTask }) {
     }
   };
 
-  //fungsi untuk mengambil data task tanpa mengubah nya
+  //fungsi untuk mengambil data task untuk menampilkan show all, done, dan todo
   const filterTask = task.filter((tasks) => tasks.id);
 
-  //Todo Search
-  // const onSearchChange = (event) => {
-  //   event.preventDefault();
-  //   setTask(event);
-  // };
-  // const search = (items, str) => {
-  //   if (str) {
-  //     return items.filter((item) => {
-  //       return item.label.toLowerCase().indexOf(str.toLowerCase()) > -1;
-  //     });
-  //   }
-  //   return items;
-  // };
-
-  // const handleSearch = (event) => {
-  //   const value = event.target.value;
-  //   setTask((item) => {
-  //     return item.filter((item) =>
-  //       item.toLowerCase().includes(value.toLowerCase)
-  //     );
-  //   });
-  // };
   //untuk merender ulang useState show dari Task
   useEffect(() => {
     setShow(task);
@@ -79,52 +39,45 @@ function LandingPage({ task, setTask }) {
     }
   };
 
-  // const handleEdite = (id) => {
-  //   const editTask = task.find((tasks) => {
-  //     return tasks.id === id;
-  //   });
-  // };
-
-  // const handleChange = (e) => {
-  //   setText(e.target.value);
-  // };
-
-  //fungsi untuk menghapus semua Task
+  //fungsi untuk menghapus semua Task yang ada
   const clearAllData = () => {
-    setShow([]);
+    const deletAll = window.confirm(
+      "Apakah Kamu Yakin Ingin Menghapus Semua Data??"
+    );
+    if (deletAll) {
+      setShow([]);
+    }
   };
 
   //Fungsi untuk menghapus task yang sudah selesai
   const clearDoneTask = () => {
-    const clearTask = task.filter((tasks) => {
-      return tasks.complete !== true;
-    });
-    setTask(clearTask);
+    const deleteDone = window.confirm(
+      "apakah kamu yakin ingin menghapus semua Task yang sudah selesai?"
+    );
+
+    if (deleteDone) {
+      const clearTask = task.filter((tasks) => {
+        return tasks.complete !== true;
+      });
+      setTask(clearTask);
+    }
   };
 
-  //Fungsi untuk menghapus show
+  //Fungsi untuk menghapus Task yang dusah di
   const deleteTask = (id) => {
-    const deleteItem = task.filter((tasks) => {
-      return id !== tasks.id;
-    });
-    setTask(deleteItem);
+    const deletes = task.find((tasks) => tasks.id === id);
+    const alertDelete = window.confirm(
+      `Apakah kamu yakin untuk menghapus: ${deletes.task}`
+    );
+    if (alertDelete) {
+      const deleteItem = task.filter((tasks) => {
+        return id !== tasks.id;
+      });
+      setTask(deleteItem);
+    }
   };
 
-  // const handleInput = (e) => {
-  //   e.preventDefault();
-
-  //   if (text.trim()) {
-  //     setTask([
-  //       ...task,
-  //       {
-  //         id: task.length + 1,
-  //         task: text,
-  //         complete: false,
-  //       },
-  //     ]);
-  //   }
-  // };
-
+  //Fungsi untuk mengatur compelete (true false) atua checkbox complete
   const handleToggle = (id) => {
     setTask((task) =>
       task.map((tasks) =>
@@ -133,43 +86,22 @@ function LandingPage({ task, setTask }) {
     );
   };
 
-  // const handleAll = () => {
-  //   window.location.reload(true);
-  // };
-
-  // const handleDone = () => {
-  //   setTask(task.filter((tasks) => tasks.complete == true));
-  // };
-  // const handleTodo = () => {
-  //   setTask(task.filter((tasks) => tasks.complete == false));
-  // };
-
   return (
     <>
       <div className=" mx-auto px-10 w-full ">
-        {/* <div className="mt-4 w-full">
-          <ToDoInput
-            onAddTask={handleInput}
-            onChange={handleChange}
-            setText={setText}
-          />
-        </div> */}
         <div className="mt-10">
-          <ToDoSearch />
+          <ToDoSearch setSearchTask={setSearchTask} />
         </div>
         <div className="mt-10">
           <TodoList onFilterTask={handlefilterTask} />
         </div>
-
         <TodoForm
-          // search={search}
           onDeleteTask={deleteTask}
           tasks={show}
           onToggleComplate={handleToggle}
           onEditeTodo={handleEdite}
-          filterTask={filterTask}
+          searchTask={searchTask}
         />
-
         <div className="mt-10">
           <TodoDelete onClearAll={clearAllData} onClearDone={clearDoneTask} />
         </div>
@@ -178,10 +110,7 @@ function LandingPage({ task, setTask }) {
   );
 }
 LandingPage.propTypes = {
-  // text: PropTypes.string,
-  task: PropTypes.func,
-  // setText: PropTypes.string,
+  task: PropTypes.array,
   setTask: PropTypes.func,
 };
-
 export default LandingPage;
